feat: add tx max gas price field (#25)
* feat: add tx max gas price field

This commit introduces an optional `max_gas_price` parameter for
transactions. When set, it will impose a cap on the permissible gas
price for a transaction's execution. If the mechanism is activated, it
results in a `Skipped` transaction effect.

This feature empowers users to restrict their gas expenditures during
unexpected gas price surges caused by market demand fluctuations.

* fix sig routine

* fix readme

* add demo-stf max gas price test