feat: add version tracking to sov-cli wallet state (#94)
This commit introduces a version tracking for the wallet state. Upon
serialization, the library generating the state will be identified by
its version number. During wallet loading, this version is validated.
If, according to SemVer, a breaking change has occurred, the wallet will
not load successfully, and the user will be presented with several
possible solutions.

An alternative approach would have been to employ reflection for
serializing the wallet state data layout. However, considering the
problem's inherent simplicity — as this library does not bind the wallet
state concretely to a context — we opted against this choice to avoid
complex derive macro implementations.

Closes https://github.com/Sovereign-Labs/sovereign-sdk/issues/1236