Preston/update risc0 fix (#1035)
* chore(repo): update risc0 version

* fix(various): update code to confirm with latest risc0 version where applicable

* chore(Cargo.toml): update rev of `risc0-cycle-utils` in that

* chore(various): rm now absent `bench` feature from `risc0-cycle-macros` deps

* fix(stf-blueprint): fix missing deps after rebase

* Allow BSL-1.0 licenses

Signed-off-by: Filippo Costa <filippo@sovlabs.io>

* fix(linting) <- update lock files after that

* Fixes for crash after upgrade to 1.0.3

This patch does 2 things:
1. drops the registration of the io_callback that wraps SYS_CYCLE_COUNT
2. adds a bincode feature flag that enables the use of bincode for host/guest serialization

The size of the counter for SYS_CYCLE_COUNT was updated to use `u64` rather than `u32`. This appears to be the source of the strange crashes seen when using the `bench` feature flag. The overriding of the SYS_CYCLE_COUNT syscall didn't seem to be necessary and it seems that metrics gathering still works properly after this change.

The `bincode` codec appears to perform 4x better in terms of cycles compared to the native `risc0` encoding.

* Initialize storage before fetching data over network

* fmt

---------

Signed-off-by: Filippo Costa <filippo@sovlabs.io>
Co-authored-by: gskapka <gregkapka@gmail.com>
Co-authored-by: Filippo Costa <filippo@sovlabs.io>
Co-authored-by: Frank Laub <frank@risczero.com>