Refactor batch builder trait to support soft confirmations. (#1511)
* Refactor batch builder trait to support soft confirmations.

This PR makes several changes. The most consequential is a significant
overhaul of the `BatchBuilder` trait to make it suitable for soft
confirmations. In addition to changing the trait definition, we move it
from sov-rollup-interface to sov-sequencer so that it can become Runtime
aware.

Next, we implement the new BatchBuilder trait for the standard
sequencer. This requires the addition of a new `SequencerDb` type for
storing the sequencer, and causes cascading changes throughout the
codebase.

* clippy

* fix docs

* fix features

* Change inner type of TxWithHash to FullyBakedTx