Adds schema-safe string (#1801)
* Replace usages of `String` with `SafeString`.

Adds a string wrapper to sov-universal-wallet that can be used to
enforce restrictions on the strings that make up CallMessages (or
schemas in general), to ensure they are safe to display to users in a
non-confusing way. Initial implementation enforces printable ASCII
characters and a configurable character limit. Constructed through
failible conversion from a `String`.