Implement chain root verification based on schema hash (#1785)
Use the schema's root hash generation to create the chain hash for demo-rollup.

Add the chain hash as a trait const on the `Runtime` trait. Use this for verification of transactions when using the normal sov authenticator (i.e. non-EVM), by concatenating the hash bytes to the encoded transaction bytes before signature generation and verification. Add support to sov-cli for generating correct signatures for this.

Test runtimes use a hash value; the demo-rollup runtime uses the real schema-generated value, written to and then read from a file. This is because `demo-rollup`'s build.rs is what generates the demo rollup's schema, but `demo-stf`, its dependency, needs to access the generated value.