Add call type generic to tx structs (#1797)
Adds a new generic to `Transaction` and `UnsignedTransaction` objects, bounded by a new trait that needs to implement a `Call` type. Blanket-implements the trait for `DispatchCall` types, setting the `Decodable` as the `Call` - which means a struct implementing `Runtime` can be used as the generic, and its `RuntimeCall` will be the call type, thus adding strong type safety to transactions. The trait allows flexibility to also implement it on e.g. bytevecs if desired in the future.

Adds a significant amount of changes through the SDK to accomodate the new typing. A large amount of testing changes stem from the adjustments to `sov-test-utils`.

---------

Co-authored-by: Theodore Bugnet <theodore@sovlabs.io>