Add standard transaction templating feature to the universal schema (#1989)
Adds transaction template generation capabilities to the schema. This means that callmessage types can be annotated with an attribute describing how they should be included in a named template, and then schema construction generates a partially pre-encoded transaction with input bindings for the remaining types that can be filled in on-demand to construct a complete transaction.

For example, a "transfer" template might accept a JSON string `{ "to": <some address>, "amount": 1000 }` and output a full, borsh-encoded RuntimeCall containing a Bank callmessage with the appropriate Transfer struct containing the above data, ready to be signed as the runtimecall of a transaction.