[Refactor]: Simplify gas metering generics (#2184)
* [Refactor]: Simplify gas metering generics

This commit removes the `GU` generic from the `GasMeter<GU>` trait and
replaces it with an associated `Spec` type. This allows simplification
of generics in many places, as well as making it possible to blanket
implement other traits in terms of the `GasMeter` trait. This was
previously impossible due to an unconstrained GU in most cases.

This new ability to use blanket impls will allow many of our state
accessor macros to be replaced with blanket impls in a future PR.

* Update trybuild output

* Fmt