Bug fix: restrict archival state to what is present in storage (#2245)
* Bug fix: restrict archival state to what is present in storage

Previously, we allowed archival queries for any existing rollup height,
even if that height was only soft confirmed. Unfortunately, this had
incoherent semantics when querying for a height that was not
the current height but was also not old enough to have been processed by
the node. When querying for those heights, users could see state from
the newer soft-confirmed heights, and possibly also from new heights
that had been fully confirmed due to incorrect setting of the
StorageVersion to query in the ApiStateAccessor.

This PR fixes the issue by making it impossible to run archival
queries for rollup heights that are not yet present in the node.

* Handle edge case of latest available version

* fmt. fix doc comment

* Update changelog. Fix doc comment

* Address code review feedback