Remove ValidityConditions (#2333)
* Remove ValidityConditions

Validity Conditions were originally added in order to support rollups
on top of Avail. They were needed because Avail commits to its data
with KZG, and we couldn't verify that commitment in zero knowledge -
so the rollup had to attach a receipt of all the inputs it had used
to be verified out-of-band by the light client.

At the current time, this no longer makes sense. First, cryptography
has advanced to the point where verifying KZG commitments is feasible -
at least for optimistic rollups. Second, the avail team never finished
implementing their adapter, and it has gotten so far out of date as
to be unusable. Finally, rollups which used validity conditions would
be unable to bridge with other chains, severely restricting their
usefullness.

Since validity conditions were never actually used in practice, it
is overwhelmingly likely that our implementation was buggy and it
added significant complexity. This commit removes validity
conditions from the SDK as an atomic change. If we wish to add them
back in future, that can be done by reverting this single commit.

* Update changelog

* lint

* Fix test compilation

* fix import

* Fix docs

* Revert "Fix docs"

This reverts commit 04c145a4b42614b72833fa0c5a829676279e59e4.