Add fixed-point decimals support to the Amount in the `Coins` struct in sov-bank (#2574)
Adds token_decimals to tokens in the bank. It's set optional in the CallMessage, defaulting to a hardcoded value, to encourage users to adhere to the ecosystem standard. 8 was chosen as the default as a reasonable amount - Bitcoin uses 8, USDC uses 6, Solana tokens tend to use between 6 and 9. Ethereum's 18 standard is very much an outlier (and usually leads to rather poor UX when displayed).

The decimal places are added to the hash when creating the TokenID. Additionally, however, the last byte of the ID is overwritten with the decimal places byte, so the actual hash is actually only 31 bytes now. This allows wallets to display amounts correctly by simply reading token_id[31] and using that to format the number. The universal wallet attributes are set up accordingly.