Sequencer update state refactor and recovery implementation (#3039)
Refactors the preferred sequencer's `update_state` to use its own managed task, which holds the state update channel, rather than the `loop_call_update_state()` mechanism. This allows the preferred sequencer to internally loop while querying for state updates - for example, it is now possible to do things like wait for the node to resync, rather than keeping track of our state across successive `update_state()` calls and trying to ensure the correct code path is taken every time.

In particular, this is most useful for

* Resyncing to the chain tip - normally the node is allowed some slack value, but if e.g. we detect some in-flight blobs we don't recognise we want to resync to the actual tip to ensure we have the latest state. This is important if the preferred sequencer's database was deleted.
* The recovery mechanism - the original motivating change for this refactor, as recovering has multiple "wait for the node to do X" conditions that introduced nested complexity under the old model.

Other than the resyncing and recovery improvements, a notable behaviour chage is that on startup, the sequencer no longer accepts transactions until it receives an update from the node. This ensures any edge cases or race conditions caused by using old node state from the ledger_db are eliminated (e.g. it is no longer possible for there to be a bug where the node is down, falls out of sync with the DA, is restarted and then the sequencer accidentally accepts some transactions before entering the resync state). However it required adjustment of some test scenarios which previously assumed the sequencer would be instantly available after a restart.

---------

Co-authored-by: Preston Evans <preston.b.evans@gmail.com>