Add solana offchain message authenticator (#1522)
Adds an authenticator module for solana offchain signed messages, and integration tests verifying transaction acceptance with a TestRollup. The transactions are JSON-serialized in order to be user-readable when displayed in the wallet, which is imperfect but functions well enough given that we can't have custom display logic and that it needs to be reversibly serializable.
Solana defines a special header (preamble) for offchain messages, which is what the Ledger app requires, but apparently not all wallets follow it so we support both spec-compliant messages (with the preamble) as well as "raw" signed messages, i.e. where the wallet just signs the provided JSON string and nothing more, nothing less.